+++ src/main/java/com/tu/model/activities/ActivityVO.java
@@ -1,5 +1,6 @@
 package com.tu.model.activities;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class ActivityVO implements java.io.Serializable {
@@ -9,7 +10,7 @@ public class ActivityVO implements java.io.Serializable {
 	private Integer actId;
 	private Integer groupId;
 	private String actName;
-	private Float actMoney;
+	private BigDecimal actMoney;
 	private String description;
 	private Date enrollStartDate;
 	private Date enrollEndDate;
@@ -47,11 +48,11 @@ public class ActivityVO implements java.io.Serializable {
 		this.actName = actName;
 	}
 
-	public Float getActMoney() {
+	public BigDecimal getActMoney() {
 		return this.actMoney;
 	}
 
-	public void setActMoney(Float actMoney) {
+	public void setActMoney(BigDecimal actMoney) {
 		this.actMoney = actMoney;
 	}
